Use Config amd Global from daemon.
