+++ WalletWasabi.Fluent/ViewModels/Settings/SettingsTabViewModelBase.cs
@@ -1,5 +1,6 @@
 using System.Reactive.Concurrency;
 using ReactiveUI;
+using WalletWasabi.Daemon;
 using WalletWasabi.Fluent.Models;
 using WalletWasabi.Fluent.ViewModels.Navigation;
 using WalletWasabi.Logging;
