+++ WalletWasabi.Fluent/WalletWasabi.Fluent.csproj
@@ -30,6 +30,7 @@
 		<PackageReference Include="System.Runtime" Version="4.3.1" />
 	</ItemGroup>
 	<ItemGroup>
+		<ProjectReference Include="..\WalletWasabi.Daemon\WalletWasabi.Daemon.csproj" />
 		<ProjectReference Include="..\WalletWasabi\WalletWasabi.csproj" />
 
 		<!-- See explanation of the additional properties here: https://github.com/dotnet/roslyn/issues/60744#issuecomment-1438273464 -->
